+++ src/Midia.java
@@ -1,6 +1,6 @@
 public interface Midia {
     void exibirInformacoes();
-    void avaliar(int nota);
+    void avaliar(double nota);
     double getMediaAvaliacoes();
     String stringBancoDados();
     void adicionarEpisodio(EpisodioSerie episodio);
